+++ packages/variance/src/props/background.ts
