+++ packages/variance/src/props/borders.ts
