+++ packages/variance/src/types/config.ts
