+++ packages/variance/src/types/properties.ts
